# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zapo Manager API
  version: 1.0.0
  description: |
    API REST para gerenciar instâncias WhatsApp via zapo-js.
    Autenticação via header `apikey` (GLOBAL_API_KEY ou chave da instância).

    **Try it out:** Configure o Server URL para sua VPS e clique em Authorize para inserir o GLOBAL_API_KEY.

    ### Formato de número
    Use DDI + DDD + número sem `+` e sem espaços: `{{number}}`.
    Números brasileiros **não** precisam do 9 extra no DDD: `555599703107` (correto), não `5555999703107`.
    Para grupos use o ID do grupo: `120363XXXXXXXX@g.us`.
servers:
  - url: http://localhost:8080
    description: Local dev
  - url: https://zapo.dominio.com
    description: VPS produção (substituir pelo domínio real)
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: apikey
      description: GLOBAL_API_KEY (aceito em todos os endpoints) ou instanceToken (endpoints da instância)
  parameters:
    instanceName:
      in: path
      name: instanceName
      required: true
      schema:
        type: string
      example: Teste-mobile
      description: Nome único da instância
...
  responses:
    Unauthorized:
      description: API key ausente ou inválida
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Unauthorized: Invalid Global API Key"
    NotFound:
      description: Instância não encontrada
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Instance not found
security:
  - ApiKeyAuth: []
...
paths:
...
  /chat/findChats/{instanceName}:
    post:
      tags: [Chat]
      summary: Buscar chats
      parameters:
        - $ref: '#/components/parameters/instanceName'
      requestBody:
        required: false
        content:
          application/json:
            example:
              where:
                jid: "{{number}}@s.whatsapp.net"
            schema:
              type: object
              properties:
                where:
                  type: object
                  description: 'Filtros Prisma (ex: { jid: "5511@s.whatsapp.net" })'
      responses:
        '200':
          description: Lista de chats
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /chat/findMessages/{instanceName}:
    post:
      tags: [Chat]
      summary: Buscar mensagens
      parameters:
        - $ref: '#/components/parameters/instanceName'
      requestBody:
        required: false
        content:
          application/json:
            example:
              where:
                jid: "{{number}}@s.whatsapp.net"
              take: 50
              skip: 0
            schema:
              type: object
              properties:
                where:
                  type: object
                  description: 'Filtros Prisma (ex: { jid: "5511@s.whatsapp.net" })'
                take:
                  type: integer
                  description: Número máximo de mensagens
                  example: 50
                skip:
                  type: integer
                  example: 0
      responses:
        '200':
          description: Lista de mensagens
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /chat/{instanceName}/{remoteJid}/status:
    get:
      tags: [Chat]
      summary: Consultar status de handoff da conversa
      description: |
        Retorna o status de controle bot/humano (pending|open|resolved) e o atendente
        atribuído. Agentes de IA devem consultar antes de enviar mensagem automática.
      parameters:
        - $ref: '#/components/parameters/instanceName'
        - name: remoteJid
          in: path
          required: true
          schema:
            type: string
          example: "5511999999999@s.whatsapp.net"
      responses:
        '200':
          description: Status atual da conversa
          content:
            application/json:
              schema:
                type: object
                properties:
                  instanceName:
                    type: string
                  remoteJid:
                    type: string
                  status:
                    type: string
                    enum: [pending, open, resolved]
                  assignedUserId:
                    type: string
                    nullable: true
                  statusChangedAt:
                    type: string
                    format: date-time
                    nullable: true
                  statusChangedBy:
                    type: string
                    nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Chat]
      summary: Alterar status de handoff da conversa
      description: |
        Superfície única de escrita reaproveitada por UI, webhooks externos e MCP tools.
        Bot só pode enviar mensagem quando status = pending; humano assumindo deve
        transicionar para open (com autoAssign opcional).
      parameters:
        - $ref: '#/components/parameters/instanceName'
        - name: remoteJid
          in: path
          required: true
          schema:
            type: string
          example: "5511999999999@s.whatsapp.net"
      requestBody:
        required: true
        content:
          application/json:
            example:
              status: "open"
              actor:
                type: "human"
                id: "luiz"
              autoAssign: true
            schema:
              type: object
              required: [status, actor]
              properties:
                status:
                  type: string
                  enum: [pending, open, resolved]
                actor:
                  type: object
                  required: [type, id]
                  properties:
                    type:
                      type: string
                      enum: [human, agent, webhook]
                    id:
                      type: string
                autoAssign:
                  type: boolean
                  description: Se true e actor.type=human e status=open, atribui automaticamente ao actor.id
      responses:
        '200':
          description: Status atualizado
        '400':
          description: status ou actor inválido
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /chat/{instanceName}/{remoteJid}/lead:
    get:
      tags: [Chat]
      summary: Obter campos do CRM (Lead) do contato
      description: |
        Retorna os valores de CRM do contato. Por padrão resolvidos com o label
        (`{"Empresa": "Acme Ltda"}`). Use `?raw=true` para obter os valores brutos
        por `slotKey` (`{"field01": "Acme Ltda"}`) — usado por formulários editáveis.
      parameters:
        - $ref: '#/components/parameters/instanceName'
        - name: remoteJid
          in: path
          required: true
          schema:
            type: string
          example: "5511999999999@s.whatsapp.net"
        - name: raw
          in: query
          required: false
          schema:
            type: boolean
          description: Se `true`, retorna valores por slotKey em vez de label
      responses:
        '200':
          description: Campos do contato (resolvidos por label ou raw por slotKey)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Chat]
      summary: Atualizar campos do CRM (Lead) do contato
      description: |
        Edita valores de CRM do contato (equivalente ao `editLead` da Uazapi).
        Chaves em `fields` devem ser `slotKey`s já mapeados via `fields-map` —
        slot não mapeado retorna 400.
      parameters:
        - $ref: '#/components/parameters/instanceName'
        - name: remoteJid
          in: path
          required: true
          schema:
            type: string
          example: "5511999999999@s.whatsapp.net"
      requestBody:
        required: true
        content:
          application/json:
            example:
              fields:
                field01: "Acme Ltda"
                field02: "5000"
              actor:
                type: "human"
                id: "luiz"
            schema:
              type: object
              required: [fields, actor]
              properties:
                fields:
                  type: object
                  additionalProperties: true
                  description: 'Pares slotKey -> valor'
                actor:
                  type: object
                  required: [type, id]
                  properties:
                    type:
                      type: string
                      enum: [human, agent, webhook]
                    id:
                      type: string
      responses:
        '200':
          description: Campos atualizados, resolvidos por label
        '400':
          description: Slot não mapeado na instância
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...